requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via django
async-timeout==4.0.2
    # via redis
backports-zoneinfo==0.2.1
    # via
    #   django
    #   django-celery-beat
    #   django-timezone-field
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via django-celery-beat
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
dj-database-url==1.2.0
    # via -r requirements.in
django==4.1.7
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-celery-beat
    #   django-redis
    #   djangorestframework
django-celery-beat==2.4.0
    # via -r requirements.in
django-environ==0.10.0
    # via -r requirements.in
django-redis==5.2.0
    # via -r requirements.in
django-timezone-field==5.0
    # via django-celery-beat
djangorestframework==3.14.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
kombu==5.2.4
    # via celery
prompt-toolkit==3.0.38
    # via click-repl
psycopg2-binary==2.9.5
    # via -r requirements.in
python-crontab==2.7.1
    # via django-celery-beat
python-dateutil==2.8.2
    # via python-crontab
pytz==2022.7.1
    # via
    #   celery
    #   django-timezone-field
    #   djangorestframework
redis==4.5.1
    # via django-redis
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
sqlparse==0.4.3
    # via django
tzdata==2022.7
    # via django-celery-beat
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
whitenoise==6.4.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
